applied patches sync-mode-names,item-type,64bit-errors,synthesis,single-tag. modified Credential LocName issue, Modified FileData and FolderData formatter
git-svn-id: https://core.forge.funambol.org/svn/core/top-level/trunk/3x/client-api/native@9856 e8e8ed6c-164c-0410-afcf-9e9a7c7d8c10

+++ src/c++/common/spds/EmailData.cpp
@@ -37,7 +37,7 @@
 
 static inline bool checkFlag(const char *xml, const char *field)
 {
-    size_t start = 0, end = 0;
+    unsigned int start = 0, end = 0;
     bool ret = false;
 
     if( XMLProcessor::getElementContent(xml, field, NULL, &start, &end) ) {
@@ -69,7 +69,7 @@ EmailData::EmailData()
 int EmailData::parse(const BCHAR *msg, size_t len)
 {
     int ret = 0;
-    size_t start, end;
+    unsigned int start, end;
     
     // Get attributes
     read      = checkFlag(msg, EMAIL_READ);
